+++ src/app/app.module.ts
@@ -2,23 +2,23 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ForgotComponent } from './forgot/forgot.component';
+import { ForgotComponent } from './features/auth/forgot/forgot.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import { ResetComponent} from './reset/reset.component';
+import { ResetComponent} from './features/auth/reset/reset.component';
 import {LoginComponent} from './features/auth/login/login.component';
 import {RegisterComponent} from './features/auth/register/register.component';
 import {HttpClientModule} from '@angular/common/http';
-import {AuthService} from './auth.service';
+import {AuthService} from './features/auth/auth.service';
 import {HeaderComponent} from './features/header/header.component';
 import {FooterComponent} from './features/footer/footer.component';
 import {NotFoundComponent} from "./features/others/notfound/not-found.component";
 import {HomePageComponent} from './features/home-page/home-page.component';
 import {ProfileComponent} from './features/profile/profile.component';
-import {EditProfileComponent} from './features/edit-profile/edit-profile.component';
-import {AuthGuard} from "./auth.guard";
+import {EditProfileComponent} from './features/profile/edit-profile/edit-profile.component';
+import {AuthGuard} from "./core/guard/auth.guard";
 import {RouterModule, Routes} from "@angular/router";
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
-import {ChangePasswordComponent} from './features/change-password/change-password.component';
+import {ChangePasswordComponent} from './features/auth/change-password/change-password.component';
 import {BuilderComponent} from './features/builder/builder.component';
 import {HeadComponent} from './features/builder/head/head.component';
 import {GalleryComponent} from './features/builder/gallery/gallery.component';
