+++ src/app/features/auth/auth.service.ts
@@ -2,7 +2,7 @@ import {Injectable, Inject} from '@angular/core';
 import {Observable} from 'rxjs';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {CanActivate, Router} from "@angular/router";
-import {LoginComponent} from "./features/auth/login/login.component";
+import {LoginComponent} from "./login/login.component";
 
 
 @Injectable()
@@ -26,22 +26,5 @@ export class AuthService {
 
   public reset(user: User): Observable<any> {
     return this.http.post(`http://localhost:8080/api/reset`, user);
-  }
-
-  public getProfile(): Observable<any> {
-    return this.http.get('http://localhost:8080/api/me', {headers: this.jwt()});
-  }
-
-  public editProfile(user: any): Observable<any> {
-    return this.http.post('http://localhost:8080/api/me', user, {headers: this.jwt()});
-  }
-
-  // register token header
-  private jwt() {
-    // register authorization header with jwt token
-    const token = localStorage.getItem('token');
-    if (token) {
-      return new HttpHeaders().set('Authorization', 'Bearer ' + token);
-    }
   }
 }
