+++ src/app/features/auth/change-password/change-password.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, ValidationErrors, ValidatorFn, Validators} from "@angular/forms";
-import {AuthService} from "../../auth.service";
+import {AuthService} from "../auth.service";
 import {Router} from "@angular/router";
