+++ src/app/features/auth/login/login.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {AuthService} from '../../../auth.service';
+import {AuthService} from '../auth.service';
 import {Router} from "@angular/router";
 import {ProfileComponent} from "../../profile/profile.component";
 // import {
