+++ src/app/features/auth/register/register.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, ValidationErrors, ValidatorFn, Validators} from '@angular/forms';
-import {AuthService} from '../../../auth.service';
+import {AuthService} from '../auth.service';
 
 @Component({
   selector: 'app-register',
