+++ src/app/features/profile/profile.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ProfileService } from './profile.service';
+
+describe('ProfileService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: ProfileService = TestBed.get(ProfileService);
+    expect(service).toBeTruthy();
+  });
+});
